Extracted from inline to build-log
