+++ build-logic/src/main/kotlin/kast.intellij-build-helpers.gradle.kts
@@ -1,0 +1,2 @@
+// Intentionally empty: applying this plugin makes build-logic task types available to
+// modules that do not otherwise use a build-logic convention plugin.
